+++ tests/admin/contact/verification/setup_utils.h
@@ -73,7 +73,7 @@ class DummyTestReturning: public Admin::ContactVerification::Test {
                 break;
             }
         }
-        T_run_result run(long _history_id) const {
+        T_run_result run(unsigned long long _history_id) const {
             return make_result(return_status, return_status);
         }
         static std::string registration_name() { return "DummyTestReturning"; }
@@ -116,7 +116,7 @@ class DummyThrowingTest: public Admin::ContactVerification::Test {
             id_ = static_cast<long>(res[0]["id_"]);
         }
 
-        T_run_result run(long _history_id) const {
+        T_run_result run(unsigned long long _history_id) const {
             throw std::runtime_error("not exactly a feature");
         }
         static std::string registration_name() { return "DummyThrowingTest"; }
